% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcms}
\alias{lcms}
\title{lcms data object}
\format{
A data frame with 11.9 million rows and 4 variables:
\describe{
  \item{mz}{Mass to charge ratio}
  \item{count}{Detector count}
  \item{scanId}{Detector scan ID}
  \item{scantime}{Scan time associated with detector scan ID}
}
}
\usage{
lcms
}
\description{
UPLC-MS data for metabolic profiling
}
\details{
A dataset containing raw UPLC-MS level 1 data.
Data were acquired using hydrophilic interaction liquid chromatography (HILIC) coupled to electrospray (ESI) quadrupole - time of flight (Q-TOF) mass spectrometer equipped with a multi-channel plate detector. ESI was performed in positive ionisation mode.
}
\keyword{datasets}
