% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_xcms}
\alias{raw_xcms}
\title{hilicPosUrine data object}
\format{
xcms object of a urine sample acquired with HILIC ESI+ Q-TOF MS:
}
\usage{
raw_xcms
}
\description{
Data were read-in using the function `xcmsRaw('/inst/extdata/Urine_HILIC_ESIpos_msLevel1.mzML', profstep = 0, includeMSn = FALSE, mslevel = 1)`
}
\keyword{datasets}
